Moving code used only by artin into its folder
