+++ lmfdb/artin_representations/cyc_alt_res_engine.py
