+++ lmfdb/artin_representations/main.py
@@ -12,7 +12,7 @@ from lmfdb.artin_representations import artin_representations_page
 from lmfdb.utils import to_dict, random_object_from_collection
 from lmfdb.search_parsing import parse_primes, parse_restricted, parse_galgrp, parse_ints, parse_count, parse_start, clean_input
 
-from lmfdb.math_classes import ArtinRepresentation
+from math_classes import ArtinRepresentation
 from lmfdb.transitive_group import group_display_knowl
 
 from sage.all import ZZ
